+++ CHANGELOG.md
@@ -25,6 +25,8 @@ See for more information about XPrivacy 3 [this FAQ](https://github.com/M66B/XPr
 
 **Next release**
 
+* Strip IP address from IP address / domain name pair for better wildcards ([issue](/../../issues/2014))
+
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
 
 **Version 3.3.3 BETA**